requirements.txt
accelerate
bitsandbytes
datasets
deepspeed==0.15.0
dm-launchpad[tensorflow]
einops
httpx
jsonlines
llm-blender==0.0.2
loralib
ml_collections
mosec
msgspec
optimum
packaging
peft
pyarrow>=0.6.1,<12
python-Levenshtein==0.25.1
scipy
torchmetrics
tqdm
transformers==4.45.2
transformers_stream_generator
vllm==0.6.2
wandb
wheel